src/Demo001_LightPluginDemo/python/requirements.txt
semantic-kernel>=1.0.0
python-dotenv>=1.0.0
